parliscope/pyproject.toml
[project]
name = "parliscope"
version = "1.0.0"
description = "Django web application for municipal transparency platform"
requires-python = ">=3.13"

dependencies = [
    "asgiref==3.8.1",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.3",
    "Django==5.1.4",
    "django-appconf==1.1.0",
    "django-cron==0.6.0",
    "django-crontab==0.7.1",
    "django-environ==0.11.2",
    "fontawesomefree==6.5.1",
    "idna==3.10",
    "psycopg2==2.9.10",
    "pysolr==3.9.0",
    "rcssmin==1.2.1",
    "requests==2.32.4",
    "rjsmin==1.2.4",
    "six==1.16.0",
    "sqlparse==0.5.3",
    "tornado==6.5.2",
    "tzdata==2025.1",
    "urllib3==2.5.0",
    "watchdog==6.0.0",
    "websocket-client==1.8.0",
    "gunicorn==23.0.0",
    "tika==2.6.0",
    "django-health-check~=3.18.3",
    "pytz~=2025.2",
    "whitenoise~=6.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest~=7.4.3",
    "pytest-env~=1.1.3",
    "pytest-django~=4.5.2",
    "ruff~=0.8.0",
    "mypy",
    "django-stubs",
    "types-requests",
]

[tool.pytest.ini_options]
testpaths = ["frontend/tests", "ris/tests"]
addopts = "--tb=short"
env = [
    "DJANGO_SETTINGS_MODULE=frontend.tests.settings"
]

[tool.ruff]
target-version = "py313"
line-length = 88
extend-exclude = ["*migrations/*.py"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["frontend", "ris"]

[tool.mypy]
python_version = "3.13"
files = ["."]
exclude = [
    "migrations/",
]
plugins = ["mypy_django_plugin.main"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "frontend.tests.settings"

[dependency-groups]
dev = [
    "django-stubs>=5.2.2",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "types-requests>=2.32.4.20250809",
]
